+++ src/main/java/com/umc/TheGoods/converter/Order/OrderConverter.java
@@ -1,25 +1,25 @@
-package com.umc.TheGoods.converter.orders;
+package com.umc.TheGoods.converter.Order;
 
 import com.umc.TheGoods.domain.enums.OrderStatus;
 import com.umc.TheGoods.domain.order.OrderDetail;
 import com.umc.TheGoods.domain.order.Orders;
 import com.umc.TheGoods.domain.types.PayType;
-import com.umc.TheGoods.web.dto.order.OrderRequest;
-import com.umc.TheGoods.web.dto.order.OrderResponse;
+import com.umc.TheGoods.web.dto.Order.OrderRequestDTO;
+import com.umc.TheGoods.web.dto.Order.OrderResponseDTO;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 public class OrderConverter {
 
-    public static OrderResponse.OrderAddResultDto toOrderAddResultDto(Orders order) {
-        return OrderResponse.OrderAddResultDto.builder()
+    public static OrderResponseDTO.OrderAddResultDto toOrderAddResultDto(Orders order) {
+        return OrderResponseDTO.OrderAddResultDto.builder()
                 .orderId(order.getId())
                 .createdAt(LocalDateTime.now())
                 .build();
     }
 
-    public static Orders toOrders(OrderRequest.OrderAddDto request) {
+    public static Orders toOrders(OrderRequestDTO.OrderAddDto request) {
         PayType payType = null;
         switch (request.getPayType()) {
             case "CARD":
@@ -44,7 +44,7 @@ public class OrderConverter {
                 .build();
     }
 
-    public static OrderDetail toOrderDetail(OrderRequest.OrderItemDto orderItemDto, Long price) {
+    public static OrderDetail toOrderDetail(OrderRequestDTO.OrderItemDto orderItemDto, Long price) {
         return OrderDetail.builder()
                 .amount(orderItemDto.getAmount())
                 .orderPrice(price)
