+++ src/main/java/com/umc/TheGoods/repository/Item/ItemRepository.java
@@ -1,4 +1,4 @@
-package com.umc.TheGoods.repository.item;
+package com.umc.TheGoods.repository.Item;
 
 import com.umc.TheGoods.domain.item.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
