+++ src/main/java/com/umc/TheGoods/service/OrderService/OrderCommandService.java
@@ -2,9 +2,9 @@ package com.umc.TheGoods.service.OrderService;
 
 import com.umc.TheGoods.domain.member.Member;
 import com.umc.TheGoods.domain.order.Orders;
-import com.umc.TheGoods.web.dto.order.OrderRequest;
+import com.umc.TheGoods.web.dto.Order.OrderRequestDTO;
 
 public interface OrderCommandService {
 
-    public Orders create(OrderRequest.OrderAddDto request, Member member);
+    public Orders create(OrderRequestDTO.OrderAddDto request, Member member);
 }
