+++ src/main/java/com/umc/TheGoods/web/dto/Order/OrderRequestDTO.java
@@ -1,4 +1,4 @@
-package com.umc.TheGoods.web.dto.order;
+package com.umc.TheGoods.web.dto.Order;
 
 import com.umc.TheGoods.validation.annotation.OrderAvailable;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.*;
 import java.util.List;
 
-public class OrderRequest {
+public class OrderRequestDTO {
     @Getter
     public static class OrderAddDto {
